раздробил на файлы по компонентам
